src/requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
browser_cookie3==0.12.0
progressbar2==3.53.1